Move extract_embeddings predict and preprocess to core/compliance
