+++ apps/fraud/compliance/api/src/tests/preprocess_test.py
@@ -1,5 +1,5 @@
 import pytest
-from pcpapillon.core.preprocess import prepare_features
+from pcpapillon.core.compliance.preprocess import prepare_features
 
 
 class PreprocessTest:
